requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.3
cymem==2.0.8
dataclasses-json==0.6.1
dnspython==2.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastapi==0.104.1
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-search-results==2.4.2
h11==0.14.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
langchain==0.0.330
langchain-experimental==0.0.37
langcodes==3.3.0
langsmith==0.0.57
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
openai==0.28.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
preshed==3.0.9
protobuf==4.25.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyarrow==14.0.0
pydantic==1.10.0
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
srsly==2.4.8
starlette==0.27.0
streamlit==1.28.1
streamlit-chat==0.1.1
sympy==1.12
tenacity==8.2.3
thinc==8.2.1
threadpoolctl==3.2.0
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.35.0
typer==0.9.0
typing-inspect==0.8.0
typing_extensions
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.23.2
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
weasel==0.3.3
yarl==1.9.2
zipp==3.17.0